Move utils to their folder.
