+++ tsutils/Singleton.py
