+++ tsutils/duration.py
